Move data mining scripts into separate directory, add FastAPI dependency
