+++ backend/.gitignore
@@ -1,3 +1,5 @@
 .idea
 *.env
 .env
+
+__pycache__/**
