+++ backend/data_mining/translator.py
