+++ backend/pyproject.toml
@@ -5,5 +5,6 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
+    "fastapi[standard]>=0.115.12",
     "pydantic-ai>=0.0.52",
 ]
